stage blog redirects working
