+++ cloudfront/wellcomelibrary.org/terraform/locals.tf
@@ -13,5 +13,12 @@ locals {
   # This should be set manually when a stable prod deploy is established.
   wellcome_library_passthru_arn_prod = "${local.wellcome_library_passthru_arn}:1"
 
+  wellcome_library_blog_redirect_arn        = aws_lambda_function.wellcome_library_blog_redirect.arn
+  wellcome_library_blog_redirect_latest     = aws_lambda_function.wellcome_library_blog_redirect.version
+  wellcome_library_blog_redirect_arn_latest = "${local.wellcome_library_blog_redirect_arn}:${local.wellcome_library_blog_redirect_latest}"
+  wellcome_library_blog_redirect_arn_stage  = local.wellcome_library_blog_redirect_arn_latest
+  # This should be set manually when a stable prod deploy is established.
+  wellcome_library_blog_redirect_arn_prod = local.wellcome_library_blog_redirect_arn_stage
+  
   edge_lambdas_bucket = data.terraform_remote_state.cloudfront_core.outputs.edge_lambdas_bucket
 }
