+++ cloudfront/wellcomelibrary.org/terraform/outputs.tf
@@ -2,6 +2,10 @@ output "wellcome_library_redirect_arn_latest" {
   value = local.wellcome_library_redirect_arn_latest
 }
 
+output "wellcome_library_blog_redirect_arn_latest" {
+  value = local.wellcome_library_blog_redirect_arn_latest
+}
+
 output "wellcome_library_passthru_arn_latest" {
   value = local.wellcome_library_passthru_arn_latest
 }
